Add navigation, routine screen and UI tweaks
Introduce a centralized navigation system and new routine features, plus multiple UI/UX improvements. Adds AegisNavigation, AegisBottomBar, RoutineScreen, RoutineCard, and data models (Routine, Screen). Refactors MainActivity to enable edge-to-edge and delegate navigation to AegisNavigation (removes inline NavHost and unused helper). Updates profile UI: TopBar height/Insets, imePadding, BiometricCard and GirthRow input handling (local temp state and numeric filtering), and Settings components (height as integer input). Changes UserProfile.height from Double to Int and adjusts ProfileViewModel (height flows, BMI calculation, updateHeight signature). Adds/adjusts strings for routines and units. Overall aims: add routine management, consolidate navigation, and improve input/layout behavior on small screens.

+++ app/src/main/java/com/yago/aegis/data/UserProfile.kt
@@ -4,7 +4,7 @@ data class UserProfile(
     val name: String,
     val disciplineDay: Int,
     val currentMass: String,
-    val height: Double,
+    val height: Int,
     val bodyFat: String,
     val goal: String,
     val basePhotoUri: String? = null,
